:recycle: Put status settings in own module
